Update: Increment pkgrel to 2 and change control mask types to u64 in mixer files

+++ sound/usb/mixer.h
@@ -81,9 +81,9 @@ struct usb_mixer_elem_list {
 struct usb_mixer_elem_info {
 	struct usb_mixer_elem_list head;
 	unsigned int control;	/* CS or ICN (high byte) */
-	unsigned int cmask; /* channel mask bitmap: 0 = master */
+	u64 cmask; /* channel mask bitmap: 0 = master */
 	unsigned int idx_off; /* Control index offset */
-	unsigned int ch_readonly;
+	u64 ch_readonly;
 	unsigned int master_readonly;
 	int channels;
 	int val_type;
